tbot/requirements.txt
channels
channels_redis

daphne

django
djangorestframework
djangorestframework-xml
django-constance[redis]==2.6.0
django-cors-headers
django-filter
django-jsonfield
django-macaddress
django-redis
django-multiselectfield
django-otp
django-guardian
django-rest-knox
drf-yasg==1.17.1
django-import-export
django-modelclone
django-celery-results
django-celery-beat

django-sendfile
drf-nested-routers
dry-rest-permissions
docker-compose
six
python-daemon==2.2.4
psycopg2==2.8.4
APScheduler==3.6.3
pyinotify
scapy==2.4.3
PyYAML==5.3
manuf==1.0.0
pyshark
kombu==4.6.8
psutil==5.7.0

Twisted[tls,http2]
tblib
mock
responses
